add exceptions in PaymentTypeServiceImpl and ExpenseTypeServiceImpl

+++ src/main/java/com/budgetPal/service/impl/ExpenseTypeServiceImpl.java
@@ -1,6 +1,8 @@
 package com.budgetPal.service.impl;
 
 import com.budgetPal.dto.ExpenseTypeDto;
+import com.budgetPal.exception.ExpenseTypeAlreadyExistException;
+import com.budgetPal.exception.ExpenseTypeNotFoundException;
 import com.budgetPal.mapper.ExpenseTypeMapper;
 import com.budgetPal.model.ExpenseType;
 import com.budgetPal.repository.ExpenseTypeRepository;
@@ -9,6 +11,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static com.budgetPal.utility.MessageConstants.EXPENSE_TYPE_ALREADY_EXIST_MESSAGE;
+import static com.budgetPal.utility.MessageConstants.EXPENSE_TYPE_NOT_FOUND_MESSAGE;
 
 @Service
 @RequiredArgsConstructor
@@ -20,6 +25,10 @@ public class ExpenseTypeServiceImpl implements ExpenseTypeService {
     @Override
     public ExpenseTypeDto expenseTypeRegister(ExpenseTypeDto expenseTypeDto) {
 
+        if (expenseTypeRepository.findByName(expenseTypeDto.name()).isPresent()) {
+            throw new ExpenseTypeAlreadyExistException(EXPENSE_TYPE_ALREADY_EXIST_MESSAGE);
+        }
+
         ExpenseType savedExpenseType = expenseTypeRepository.save(expenseTypeMapper.toEntity(expenseTypeDto));
 
         return expenseTypeMapper.toDto(savedExpenseType);
@@ -27,7 +36,13 @@ public class ExpenseTypeServiceImpl implements ExpenseTypeService {
 
     @Override
     public void deleteExpenseType(String name) {
-        expenseTypeRepository.deleteById(expenseTypeRepository.findByName(name).get().getExpenseTypeId());
+
+        if (expenseTypeRepository.findByName(name).isPresent()) {
+            expenseTypeRepository.deleteById(expenseTypeRepository.findByName(name).get().getExpenseTypeId());
+
+        }else {
+            throw new ExpenseTypeNotFoundException(EXPENSE_TYPE_NOT_FOUND_MESSAGE);
+        }
     }
 
     @Override
